+++ hooks/react-client/useClickOutside.ts
