+++ next.orama.mjs
@@ -1,9 +1,11 @@
 import { Highlight } from '@orama/highlight';
 import { OramaClient } from '@oramacloud/client';
 
+import { DEFAULT_ORAMA_QUERY_PARAMS } from './next.constants.mjs';
+
 export const orama = new OramaClient({
-  endpoint: process.env.NEXT_PUBLIC_ORAMA_ENDPOINT!,
-  api_key: process.env.NEXT_PUBLIC_ORAMA_API_KEY!,
+  endpoint: process.env.NEXT_PUBLIC_ORAMA_ENDPOINT,
+  api_key: process.env.NEXT_PUBLIC_ORAMA_API_KEY,
 });
 
 orama.startHeartBeat({ frequency: 3500 });
@@ -16,8 +18,6 @@ export const highlighter = new Highlight({
 export async function getInitialFacets() {
   return await orama.search({
     term: '',
-    facets: {
-      siteSection: {},
-    },
+    ...DEFAULT_ORAMA_QUERY_PARAMS,
   });
 }
